+++ bundles/framework/geometrycutter/index.js
@@ -1,0 +1,4 @@
+import './instance';
+
+// register create function for bundleid
+Oskari.bundle('geometrycutter', () => Oskari.clazz.create('Oskari.mapframework.bundle.geometrycutter.GeometryCutterBundleInstance'));