+++ bundles/framework/guidedtour/index.js
@@ -1,0 +1,4 @@
+import './instance';
+
+// register create function for bundleid
+Oskari.bundle('guidedtour', () => Oskari.clazz.create('Oskari.framework.bundle.guidedtour.GuidedTourBundleInstance'));