+++ bundles/framework/layeranalytics/index.js
@@ -1,0 +1,4 @@
+import './instance';
+
+// register create function for bundleid
+Oskari.bundle('layeranalytics', () => Oskari.clazz.create('Oskari.layeranalytics.StatusBundleInstance'));