+++ bundles/framework/mapfull/index.js
@@ -1,0 +1,4 @@
+import './instance';
+
+// register create function for bundleid
+Oskari.bundle('mapfull', () => Oskari.clazz.create('Oskari.mapframework.bundle.mapfull.MapFullBundleInstance'));