+++ bundles/framework/postprocessor/index.js
@@ -1,0 +1,4 @@
+import './instance';
+
+// register create function for bundleid
+Oskari.bundle('postprocessor', () => Oskari.clazz.create('Oskari.mapframework.bundle.postprocessor.PostProcessorBundleInstance'));