+++ bundles/framework/userguide/index.js
@@ -1,0 +1,4 @@
+import './instance';
+
+// register create function for bundleid
+Oskari.bundle('userguide', () => Oskari.clazz.create('Oskari.mapframework.bundle.userguide.UserGuideBundleInstance'));