+++ bundles/sample/myfourthbundle/resources/locale/fr.js
@@ -1,9 +1,0 @@
-Oskari.registerLocalization(
-{
-    "lang": "fr",
-    "key": "MyFourthBundle",
-    "value": {
-        "tooltip": "Click me!",
-        "toolClicked": "Tool clicked"
-    }
-});